+++ tests/test_scrape.py
@@ -45,6 +45,18 @@ def test_added_transformer_should_be_usable(shining_content):
     assert data == {"year": 2005}
 
 
+def test_transformers_should_be_chainable(shining_content):
+    transformers.century = lambda x: x // 100 + 1
+    items = [
+        {
+            "key": "century",
+            "value": {"path": '//span[@class="year"]/text()', "transform": "int|century"},
+        }
+    ]
+    data = scrape(shining_content, {"items": items})
+    assert data == {"century": 20}
+
+
 def test_unknown_transformer_should_raise_error(shining_content):
     with pytest.raises(ValueError):
         items = [